# api/openapi.yaml
openapi: 3.0.3
info:
  title: PicFast
  version: "1.0"
  description: Image hosting service API
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
security: []
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout
      security:
        - BearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/me:
    get:
      tags: [users]
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [users]
      summary: Update current user profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileUpdateResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /images:
    get:
      tags: [images]
      summary: List user images
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PaginatedImages"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [images]
      summary: Upload an image
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ImageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /upload:
    post:
      tags: [images]
      summary: Upload an image (guest)
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ImageResponse"
  /images/{key}:
    get:
      tags: [images]
      summary: Get image detail
      parameters:
        - $ref: "#/components/parameters/ImageKeyParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ImageResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [images]
      summary: Delete image
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ImageKeyParam"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      tags: [images]
      summary: Update image (permission, album)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ImageKeyParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateImageRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ImageResponse"
  /albums:
    get:
      tags: [albums]
      summary: List user albums
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PaginatedAlbums"
    post:
      tags: [albums]
      summary: Create album
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlbumRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AlbumResponse"
  /albums/{id}:
    put:
      tags: [albums]
      summary: Update album
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlbumRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AlbumUpdateResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [albums]
      summary: Delete album
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "403":
          $ref: "#/components/responses/Forbidden"
  /strategies:
    get:
      tags: [strategies]
      summary: List available strategies
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyListResponse"
  /admin/users:
    get:
      tags: [admin]
      summary: List all users
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PaginatedAdminUsers"
  /admin/users/{id}:
    get:
      tags: [admin]
      summary: Get user detail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [admin]
      summary: Update user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
    delete:
      tags: [admin]
      summary: Delete user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /admin/groups:
    get:
      tags: [admin]
      summary: List groups
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
    post:
      tags: [admin]
      summary: Create group
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Created
  /admin/groups/{id}:
    get:
      tags: [admin]
      summary: Get group detail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
    put:
      tags: [admin]
      summary: Update group
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupRequest"
      responses:
        "200":
          description: OK
    delete:
      tags: [admin]
      summary: Delete group
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /admin/groups/{id}/strategies:
    put:
      tags: [admin]
      summary: Set group strategies
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [strategy_ids]
              properties:
                strategy_ids:
                  type: array
                  items:
                    type: integer
                    format: int64
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /admin/strategies:
    get:
      tags: [admin]
      summary: List strategies
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
    post:
      tags: [admin]
      summary: Create strategy
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStrategyRequest"
      responses:
        "201":
          description: Created
  /admin/strategies/{id}:
    get:
      tags: [admin]
      summary: Get strategy
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
    put:
      tags: [admin]
      summary: Update strategy
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
    delete:
      tags: [admin]
      summary: Delete strategy
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: "#/components/responses/OK"
...
  /admin/images:
    get:
      tags: [admin]
      summary: List all images
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: OK
  /admin/images/{id}:
    delete:
      tags: [admin]
      summary: Delete image
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /admin/moderation/pending:
    get:
      tags: [admin]
      summary: List pending moderation items
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ModerationItem"
  /admin/moderation/{id}/approve:
    post:
      tags: [admin]
      summary: Approve a moderation item
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /admin/moderation/{id}/reject:
    post:
      tags: [admin]
      summary: Reject a moderation item
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /admin/settings:
    get:
      tags: [admin]
      summary: Get settings
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SettingsResponse"
    put:
      tags: [admin]
      summary: Update settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: OK
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    ImageKeyParam:
      name: key
      in: path
      required: true
      schema:
        type: string
  responses:
    OK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessResponse"
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    SuccessResponse:
      type: object
      properties:
        status:
          type: boolean
        message:
          type: string
        data:
          type: object
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int64
        token_type:
          type: string
          example: Bearer
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    ImageResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        key:
          type: string
        origin_name:
          type: string
        size_bytes:
          type: integer
          format: int64
        mimetype:
          type: string
        extension:
          type: string
        width:
          type: integer
          format: int32
        height:
          type: integer
          format: int32
        md5:
          type: string
        sha1:
          type: string
        permission:
          type: integer
          format: int16
        album_id:
          type: integer
          format: int64
          nullable: true
        links:
          $ref: "#/components/schemas/ImageLinks"
        created_at:
          type: string
          format: date-time
...
    PaginatedImages:
      allOf:
        - $ref: "#/components/schemas/PaginatedData"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ImageListItem"
    UploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        strategy_id:
          type: integer
          format: int64
        album_id:
          type: integer
          format: int64
        permission:
          type: integer
...
    UserProfileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        role:
          type: string
        status:
          type: integer
        capacity_bytes:
          type: integer
          format: int64
        used_bytes:
          type: integer
          format: int64
        image_num:
          type: integer
          format: int64
        album_num:
          type: integer
          format: int64
        settings:
          type: object
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
    UserProfileUpdateResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        settings:
          type: object
        updated_at:
          type: string
          format: date-time
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        password:
          type: string
          minLength: 8
        settings:
          type: object
...